V6

+++ src/main/java/com/mycompany/gs1_prototipo1/model/Coordinate.java
@@ -4,7 +4,7 @@ public class Coordinate {
     private double x;
     private double y;
 
-    public Coordinate(float x, float y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
